src/main/java/fashionable/simba/yanawaserver/auth/token/JwtTokenProvider.java
package fashionable.simba.yanawaserver.auth.token;

import io.jsonwebtoken.Claims;
import io.jsonwebtoken.Jws;
import io.jsonwebtoken.JwtException;
import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.SignatureAlgorithm;

import java.util.Date;
import java.util.List;

public class JwtTokenProvider {
    private static final String ROLES = "roles";
    private final String secretKey;
    private final long validityMilliseconds;

    public JwtTokenProvider(String secretKey, long validityMilliseconds) {
        this.secretKey = secretKey;
        this.validityMilliseconds = validityMilliseconds;
    }

    public String createToken(String principal, List<String> roles) {
        Claims claims = Jwts.claims().setSubject(principal);
        Date now = new Date();
        Date validity = new Date(now.getTime() + validityMilliseconds);

        return Jwts.builder()
            .setClaims(claims)
            .setIssuedAt(now)
            .setExpiration(validity)
            .claim(ROLES, roles)
            .signWith(SignatureAlgorithm.HS256, secretKey)
            .compact();
    }

    public String getPrincipal(String token) {
        return Jwts.parser().setSigningKey(secretKey).parseClaimsJws(token).getBody().getSubject();
    }

    public List<String> getRoles(String token) {
        return Jwts.parser().setSigningKey(secretKey).parseClaimsJws(token).getBody().get(ROLES, List.class);
    }

    public boolean validateToken(String token) {
        try {
            Jwts.parser().setSigningKey(secretKey).parseClaimsJws(token);
            return true;
        } catch (JwtException | IllegalArgumentException exception) {
            return false;
        }
    }
}
